requirements.txt
dateparser~=1.0.0
numpy~=1.17.0
matplotlib~=3.1.1
pandas~=0.25.0
python-dateutil~=2.8.0
scipy~=1.5.4